requirements.txt
annotated-types==0.7.0
anyio==4.9.0
Authlib==1.6.1
bidict==0.23.1
blinker==1.9.0
cachelib==0.13.0
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.5
dnspython==2.7.0
email_validator==2.2.0
eventlet==0.40.1
Flask==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-OAuthlib==0.9.6
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
google-auth==2.40.3
google-genai==1.24.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
narwhals==1.44.0
oauthlib==2.1.0
packaging==25.0
plotly==6.0.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
python-engineio==4.12.2
python-socketio==5.13.0
requests==2.32.4
requests-oauthlib==1.1.0
rsa==4.9.1
simple-websocket==1.1.0
sniffio==1.3.1
SQLAlchemy==2.0.41
tenacity==8.5.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
WTForms==3.2.1
